safari-534.53-branch/LayoutTests/fast/files/apply-blob-url-to-xhr-expected.txt

CONSOLE MESSAGE: line 29: TypeError: 'undefined' is not an object (evaluating 'window.webkitURL.createObjectURL')
FAIL: Timed out waiting for notifyDone to be called
